force unaligning equal signs and arrows
Alignment can seem like a good idea at first, but it makes diff harder
to read, and bigger, thus increasing the risk of conflicts. It also adds
a load on the developer because it requires maintenance when they add or
remove the longest key of the array / variable of the assignment list.