Set --prefer-stable option as default
For some reasons like fixing cirular dependencies, we have to add @dev dependencies

See: https://github.com/sonata-project/SonataAdminBundle/pull/3777

But if we do that, only dev dependencies will be installed for test, this is not what we want.

This is why I added --prefer-stable option on composer update process.

The next step would be to remove this option on allow_failures section if the following issue is resolved.

https://github.com/travis-ci/travis-ci/issues/6021