Use the weak vendors mode
This is stricter than the weak mode, but still will not make the build
fail if the deprecation comes from a vendor. However, if a deprecation
is contributed and we call our own deprecated code, or do so in a test
without adding said test to the "legacy" phpunit group, the build will
fail.