Stop testing unsupported versions of PHP
Regardless of whether we continue to support php 5.6 or php 7.0, if a
new bug is discovered, and if affects these versions of php, and if
affects us, we will be stuck and will not be able to move forward. So
let us just stop testing with these versions and decide whether we stop
supporting them and when separately.

See https://secure.php.net/supported-versions.php